git(add): error handling helper for mongoose validation
Errors that are not thrown because of a Mongoose validator violation will contain an
associated error code. In some cases, these errors need to be handled differently. For
example, errors caused due to a violation of the unique constraint will return an
error object that is different from Mongoose validation errors. The unique option is
not a validator but a convenient helper for building MongoDB unique indexes, so we
will add another getUniqueErrorMessage method to parse the unique constraint-related error object and construct an appropriate error message

By using the getErrorMessage function that's exported from this helper file, we can
add meaningful error messages when handling errors that are thrown by Mongoose
operations.