feat(add): user sign in controller
The POST request object receives the email and password in req.body. This email is
used to retrieve a matching user from the database. Then, the password
authentication method defined in UserSchema is used to verify the password that's
received in req.body from the client.
If the password is successfully verified, the JWT module is used to generate a signed
JWT using a secret key and the user's _id value. Then, the signed JWT is returned to the authenticated client, along with the user's
details. Optionally, we can also set the token to a cookie in the response object so that
it is available to the client-side if cookies are the chosen form of JWT storage. On the
client-side, this token must be attached as an Authorization header when
requesting protected routes from the server. To sign-out a user, the client-side can
simply delete this token depending on how it is being stored. In the next section, we
will learn how to use a signout API endpoint to clear the cookie containing the
token.